pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "maid"
version = "0.1.0"
description = "Medium-Agnostic AI Duo — a real-time AI companion for gaming"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "mss>=9.0.0",
    "sounddevice>=0.4.6",
    "numpy>=1.26.0",
    "Pillow>=10.0.0",
    "elevenlabs>=1.0.0",
    "openai>=1.30.0",
    "pyttsx3>=2.90",
]

[project.scripts]
maid = "maid.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["maid*"]
